requirements.txt
streamlit==1.38.0
python-dotenv==1.0.1
langchain==0.2.12
langchain-openai==0.1.21
pandas==2.2.2
pydantic==2.9.2
pyarrow==15.0.2
